math/Utl: add (float) cast in Limit to generate frsp before fctiwz
rndobj/AnimFilter: fix SetFrame callee-save allocation and StartFrame register order

SetFrame: remove named offset var (prevents extra callee-save), cache mStart/mEnd
after RandomFloat call to match target FPR allocation pattern (95.2% match).
StartFrame: swap declaration order (start before offset) to fix f2/f3 swap (100% match).
Utl: (int)(float)floor(...) generates frsp+fctiwz matching DC3/target vs plain floor.

+++ src/system/math/Utl.h
@@ -188,7 +188,7 @@ inline bool PowerOf2(int num) {
 
 inline float Limit(float f1, float f2, float f3, int &i) {
     float fsub = f2 - f1;
-    int floored = floor((f3 - f1) / fsub);
+    int floored = (int)(float)floor((f3 - f1) / fsub);
     i = floored;
     return -(floored * fsub - f3);
 }
